requirements.txt
streamlit==1.29.0
python-dotenv==1.0.0
pydub==0.25.1
ffmpeg-python==0.2.0
ollama==0.1.6
speechrecognition==3.10.1
pyaudio==0.2.14
